+++ packages/namehash-ui/src/components/registrar-actions/RegistrarActionCard.tsx
@@ -16,18 +16,19 @@ import {
   ZERO_ENCODED_REFERRER,
 } from "@ensnode/ensnode-sdk";
 
-import { DisplayDuration } from "@/components/datetime/DisplayDuration.tsx";
-import { RelativeTime } from "@/components/datetime/RelativeTime.tsx";
-import type { IdentityLinkDetails } from "@/components/identity/Identity.tsx";
-import { NameDisplay } from "@/components/identity/Name.tsx";
+import { useIsMobile } from "../../hooks/useIsMobile";
+import { getBlockExplorerTransactionDetailsUrl } from "../../utils/blockExplorers";
+import { cn } from "../../utils/cn";
+import { getEnsManagerAddressDetailsUrl } from "../../utils/ensManager";
+import { DisplayDuration } from "../datetime/DisplayDuration";
+import { RelativeTime } from "../datetime/RelativeTime";
+import type { IdentityLinkDetails } from "../identity/Identity";
+import { NameDisplay } from "../identity/Name";
 import {
   ResolveAndDisplayIdentity,
   type ResolveAndDisplayIdentityProps,
-} from "@/components/identity/ResolveAndDisplayIdentity.tsx";
-import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip.tsx";
-import { useIsMobile } from "@/hooks/useIsMobile.tsx";
-import { getBlockExplorerTransactionDetailsUrl } from "@/utils/blockExplorers.ts";
-import { cn } from "@/utils/cn.ts";
+} from "../identity/ResolveAndDisplayIdentity";
+import { Tooltip, TooltipContent, TooltipTrigger } from "../ui/tooltip";
 
 interface LabeledFieldProps {
   fieldLabel: string;
@@ -54,7 +55,7 @@ function LabeledField({ fieldLabel, className, children }: PropsWithChildren<Lab
 }
 
 interface ResolveAndDisplayReferrerIdentityProps
-  extends Omit<ResolveAndDisplayIdentityProps, "identity"> {
+  extends Omit<ResolveAndDisplayIdentityProps, "identity" | "identityLinkDetails"> {
   chainId: DefaultableChainId;
   referral: RegistrarActionReferral;
 }
@@ -70,7 +71,6 @@ function ResolveAndDisplayReferrerIdentity({
   referral,
   accelerate = true,
   withLink = true,
-  identityLinkDetails,
   withTooltip = true,
   withAvatar = false,
   withIdentifier = true,
@@ -144,13 +144,17 @@ function ResolveAndDisplayReferrerIdentity({
       withIdentifier={withIdentifier}
       className={className}
       withLink={withLink}
-      identityLinkDetails={identityLinkDetails}
+      identityLinkDetails={{
+        isExternal: true,
+        link: getEnsManagerAddressDetailsUrl(referrerIdentity.address, namespaceId),
+      }}
     />
   );
 }
 
 export interface RegistrarActionCardLoadingProps {
   showReferrer?: boolean;
+  showReferralProgramField?: boolean;
 }
 
 /**
@@ -158,6 +162,7 @@ export interface RegistrarActionCardLoadingProps {
  */
 export function RegistrarActionCardLoading({
   showReferrer = true,
+  showReferralProgramField = true,
 }: RegistrarActionCardLoadingProps) {
   const isMobile = useIsMobile();
 
@@ -208,9 +213,11 @@ export function RegistrarActionCardLoading({
         </div>
       )}
 
-      <LabeledField fieldLabel="Incentive program" className="nhui:w-[15%] nhui:min-w-[162px]">
-        <div className="nhui:animate-pulse nhui:h-[14px] nhui:mt-[4px] nhui:mb-[3px] nhui:bg-gray-200 nhui:rounded-sm nhui:w-1/4 nhui:sm:w-4/5" />
-      </LabeledField>
+      {showReferralProgramField && (
+        <LabeledField fieldLabel="Incentive program" className="nhui:w-[15%] nhui:min-w-[162px]">
+          <div className="nhui:animate-pulse nhui:h-[14px] nhui:mt-[4px] nhui:mb-[3px] nhui:bg-gray-200 nhui:rounded-sm nhui:w-1/4 nhui:sm:w-4/5" />
+        </LabeledField>
+      )}
     </div>
   );
 }
@@ -222,7 +229,6 @@ export interface RegistrarActionCardProps {
   links: {
     name: IdentityLinkDetails;
     registrant: IdentityLinkDetails;
-    referrer?: IdentityLinkDetails;
   };
   showReferrer?: boolean;
   referralProgramField?: ReactNode;
@@ -267,13 +273,19 @@ export function RegistrarActionCard({
     nhui:items-start nhui:gap-2 nhui:p-4 nhui:sm:p-6 nhui:sm:gap-y-5 nhui:rounded-2xl nhui:border nhui:border-gray-200 nhui:text-sm nhui:bg-white"
     >
       <LabeledField fieldLabel="Name" className="nhui:w-[15%] nhui:min-w-[162px]">
-        <a
-          target="_blank"
-          href={links.name.link.href}
-          className="nhui:max-sm:max-w-3/4 nhui:sm:w-full nhui:box-border nhui:overflow-x-auto nhui:text-blue-600 nhui:font-medium nhui:hover:underline nhui:hover:underline-offset-[25%] nhui:whitespace-nowrap"
-        >
-          <NameDisplay name={namedRegistrarAction.name} className="nhui:h-[21px]" />
-        </a>
+        {links.name.link !== null ? (
+          <a
+            target={links.name.isExternal ? "_blank" : "_self"}
+            href={links.name.link.href}
+            className="nhui:max-sm:max-w-3/4 nhui:sm:w-full nhui:box-border nhui:overflow-x-auto nhui:text-blue-600 nhui:font-medium nhui:hover:underline nhui:hover:underline-offset-[25%] nhui:whitespace-nowrap"
+          >
+            <NameDisplay name={namedRegistrarAction.name} className="nhui:h-[21px]" />
+          </a>
+        ) : (
+          <p className="nhui:max-sm:max-w-3/4 nhui:sm:w-full nhui:box-border nhui:overflow-x-auto nhui:font-medium nhui:whitespace-nowrap">
+            <NameDisplay name={namedRegistrarAction.name} className="nhui:h-[21px]" />
+          </p>
+        )}
       </LabeledField>
 
       <LabeledField
@@ -330,7 +342,6 @@ export function RegistrarActionCard({
               withAvatar={true}
               withIdentifier={false}
               withTooltip={false}
-              identityLinkDetails={links.referrer}
             />
           )}
           <LabeledField fieldLabel="Referrer" className="nhui:w-[15%] nhui:min-w-[110px]">
@@ -341,7 +352,6 @@ export function RegistrarActionCard({
               withAvatar={isMobile}
               withIdentifier={true}
               withTooltip={false}
-              identityLinkDetails={links.referrer}
             />
           </LabeledField>
         </div>