+++ packages/ensnode-schema/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ensnode/ensnode-schema
 
+## 1.5.1
+
 ## 1.5.0
 
 ### Minor Changes